requirements.txt
asgiref==3.5.2
awscli==1.24.2
beautifulsoup4==4.11.1
boto3==1.23.2
botocore==1.26.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.4
cryptography==38.0.1
defusedxml==0.7.1
Django==3.2.7
django-allauth==0.45.0
django-bootstrap5==2.1.2
django-rest-framework==0.1.0
django-widgets-improved==1.5.0
djangorestframework==3.14.0
docutils==0.15.2
idna==3.3
jmespath==1.0.0
numpy==1.23.1
oauthlib==3.2.0
opencv-python==4.6.0.66
Pillow==8.3.2
psycopg2-binary==2.9.1
pyasn1==0.4.8
pycparser==2.21
PyJWT==2.4.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.2.1
PyYAML==5.4.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.7.2
s3transfer==0.5.2
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
urllib3==1.26.12
